organization and some updates
