+++ controllers/home.go
@@ -1,0 +1,11 @@
+package controllers
+
+import (
+	"fmt"
+
+	"github.com/gin-gonic/gin"
+)
+
+func Home(c *gin.Context) {
+	fmt.Println("Here is a home")
+}
