+++ models/genre.go
@@ -1,0 +1,6 @@
+package models
+
+type Genre struct {
+	GenreId     int64  `json:"genre_id"`
+	Description string `json:"description"`
+}
