+++ models/store.go
@@ -1,0 +1,6 @@
+package models
+
+type Store struct {
+	StoreId int64  `json:"store_id"`
+	Name    string `json:"name"`
+}
